use query wrapped in suspense
