+++ app/api/contract-due-email/route.ts
@@ -7,44 +7,44 @@ export async function GET() {
   const today = new Date();
   const targetDate = new Date(today);
   targetDate.setDate(today.getDate() + 1);
+  // Fetch all contracts that are due tomorrow
   const { data, error, status } = await admin
     .from("contracts")
     .select("*")
     .eq("due_date", format(targetDate, "yyyy-MM-dd"))
     .order("created_at", { ascending: false });
 
-  if (data) {
-    const contracts: Database["contract"][] = data;
+  if (error) {
+    return Response.json(`Something went wrong, ERROR_CODE: ${status}`, {
+      status,
+    });
+  }
 
-    await Promise.all(
-      contracts.map(async (contract) => {
-        const { data: retailer } = await admin
-          .from("users")
-          .select("*")
-          .eq("id", contract.retailer_id)
-          .single<Database["user"]>();
+  // Send message to all retailers have contract due
+  const contracts: Database["contract"][] = data;
+  await Promise.all(
+    contracts.map(async (contract) => {
+      const { data: retailer } = await admin
+        .from("users")
+        .select("*")
+        .eq("id", contract.retailer_id)
+        .single<Database["user"]>();
 
-        if (!retailer) return;
+      if (!retailer) return;
 
-        await sendEmail({
-          to: retailer.email,
-          subject: "اشعار دفع",
-          html: `
+      await sendEmail({
+        to: retailer.email,
+        subject: "اشعار دفع",
+        html: `
             <h1>مرحبا, ${retailer.full_name}</h1>
             <h3>باقي يوم علي الدفع</h3>
             <p>تاريخ استحقاق الدفعه يوم: <strong>${contract.due_date}</strong></p>
             <p>مبلغ الاستحقاق: <strong>${(contract.amount / contract.number_of_payments).toLocaleString()} ر.س</strong></p>
             <p>.فريق عمل منصة العقود يتمني لك النجاح</p>
         `,
-        });
-      })
-    );
-  }
-  if (error) {
-    return Response.json(`Something went wrong, ERROR_CODE: ${status}`, {
-      status,
-    });
-  }
+      });
+    })
+  );
 
   return Response.json(`Everything is OK, CODE: ${status}`, { status });
 }
